// frontend/src/typings/Environment.ts
export interface Environment {
    acidificationPotential: number;
    ba: string;
    eutrophicationPotential: number;
    fossilFuelEnergy: number;
    frec: string;
    globalWarmingPotential: number;
    nonRenewableEnergy: number;
    ozoneDepletionPotential: number;
    particulateMatterFormationPotential: number;
    photochemicalSmogFormationPotential: number;
    primaryEnergyDemand: number;
    renewableEnergy: number;
    state: string;
    zipcode: number;
}